0.3 release

+++ src/org/nagios/JMXQuery.java
@@ -29,12 +29,12 @@ public class JMXQuery {
 	private int verbatim;
 	private JMXConnector connector;
 	private MBeanServerConnection connection;
-	private int warning, critical;
+	private long warning, critical;
 	private String attribute, info_attribute;
 	private String attribute_key, info_key;
 	private String object;
 	
-	private int checkData;
+	private long checkData;
 	private Object infoData;
 	
 	private static final int RETURN_OK = 0; // 	 The plugin was able to check the service and it appeared to be functioning properly
@@ -172,7 +172,7 @@ public class JMXQuery {
 	}
 
 
-	private boolean compare(int level, boolean more) {
+	private boolean compare(long level, boolean more) {
 		if(more)
 			return checkData>=level;
 		else
@@ -205,11 +205,11 @@ public class JMXQuery {
 		
 	}
 
-	private int parseData(Object o) {
+	private long parseData(Object o) {
 		if(o instanceof Number)
-			return ((Number)o).intValue();
+			return ((Number)o).longValue();
 		else 
-			return Integer.parseInt(o.toString());
+			return Long.parseLong(o.toString());
 	}
 
 
@@ -237,9 +237,9 @@ public class JMXQuery {
 				}else if(option.startsWith("-v")){
 					this.verbatim = option.length()-1;
 				}else if(option.equals("-w")){
-					this.warning = Integer.parseInt(args[++i]);
+					this.warning = Long.parseLong(args[++i]);
 				}else if(option.equals("-c")){
-					this.critical = Integer.parseInt(args[++i]);
+					this.critical = Long.parseLong(args[++i]);
 				}
 			}
